Add option to only add objects to collection, without unlinking

+++ collection/operator/move_selected_to_this.py
@@ -7,6 +7,7 @@ class GU_OT_collection_move_selected_to_this(bpy.types.Operator):
     bl_idname = "collection.move_selected_to_this"
     bl_label = "Move selected objects to this collection"
     bl_options = {"UNDO"}
+    unlink_others: bpy.props.BoolProperty(default=True)
 
     def execute(self, context):
         target_col = context.collection
@@ -19,7 +20,8 @@ class GU_OT_collection_move_selected_to_this(bpy.types.Operator):
                 if col == target_col:
                     already_linked = True
                 else:
-                    col.objects.unlink(obj)
+                    if self.unlink_others:
+                        col.objects.unlink(obj)
             if not already_linked:
                 target_col.objects.link(obj)
